run_model now breaks the image before detection
run_model now breaks the image before detection (according to model metadata). this approach does not require size generalization for the object detection with the model, making the model more effective by simplifying the task.

tool still does not save metadata, the current metadata was made manually (TODO)

the information of how the image was broken is already stored for future use so that it can be aggregated to form a standard detection notation even if the model performed multiple ones